+++ DTCD-WorkspaceSystem/src/TabsPanelComponent.vue
@@ -34,7 +34,7 @@
             status_editOn: edit,
             status_editName: editName
           }"
-          @click="setActiveTab(id)"
+          @click="tabClickHandler(id)"
         >
           <span class="TabName">{{ name | truncate( 21, '...') }}</span>
 
@@ -112,6 +112,9 @@ export default {
       editMode: $root.editMode,
       visibleNavBar: $root.visibleNavBar,
       tabsSwitcherInstance: $root.tabsSwitcherInstance,
+      logSystem: $root.logSystem,
+      eventSystem: $root.eventSystem,
+      currentTab: '',
       tabName: '',
       scrollButton: false,
       scrollWidth: null,
@@ -174,12 +177,25 @@ export default {
       }
       return tabId;
     },
+
+    tabClickHandler(id) {
+      const clickedTab = this.getTab(id);
+
+      this.eventSystem.publishEvent('WorkspaceTabClicked', clickedTab);
+      this.logSystem.debug(`Clicked "${id}" workspace tab`);
+
+      if (id !== this.currentTab) this.setActiveTab(id);
+    },
+
     setActiveTab(tabId) {
       this.tabsCollection.forEach((tab) => {
         tab.isActive = tab.id === tabId;
       });
       this.tabsSwitcherInstance.activeTab(tabId);
+      this.currentTab = tabId;
+      this.logSystem.debug(`Activated "${tabId}" workspace tab`);
     },
+
     removeTab(tabId, index) {
       if (!tabId || this.tabsCollection.length === 1) {
         return;
@@ -193,19 +209,23 @@ export default {
       }
 
     },
+
     getTab(tabId) {
       if (!tabId) {
         return;
       }
       return this.tabsCollection.find((tab) => tab.id === tabId);
     },
+
     getIdNewTab() {
       return `wss-tab-${Math.ceil(Math.random() * 10000)}`;
     },
+
     changeName(index, name) {
       this.tabsCollection[index].editName = true;
       this.tabName = name;
     },
+
     saveName(index) {
       const checkTabName = this.tabsCollection.find((tab, i) => tab.name === this.tabName && i !== index);
       if (checkTabName) {
@@ -215,9 +235,11 @@ export default {
       this.tabsCollection[index].editName = false;
       this.tabsCollection[index].name = this.tabName;
     },
+
     resetError(index) {
       this.$refs.input[index].invalid = false;
     },
+
     scroll(toRight) {
       this.$refs.tabList.scrollLeft += toRight
           ? (this.$refs.tabPanel.clientWidth * 0.8)